Fix PDB generation and random scaffold generation
Two separate fixes in this commit:
The gen_PDB file was updated to include hybrid base-36 encoding
for structures with more than 99,999 atoms (the limit in the
native PDB file format), so that all atoms will be properly
displayed in the PDB viewer, e.g. UCSF Chimera. Additionally,
the generation of separate PDB files (with either all strands as
separate chains, each strand as its own model, or all strands as
one chain with its own segid) was fixed so that the files are
different from each other, and the ends of chains have "TER"
lines in each PDB file, which is required for some downstream
programs.

When no scaffold sequence is provided as input, the code will either
use M13 scaffold sequence or, for sufficiently large structures,
randomly generate a sequence to use. The program incorrectly handled
lowercase letters in scaffold sequence, from the randomly generated
sequence or from an input sequence file. This fix allows the code
to read both upper and lowercase letters, and provides an error msg
for unrecognized nucleotides (not A,C,T,G, or U).