Update scripts to Python 3 and Networkx >= 3.0
All scripts were updated from Python 2 (now dead) to Python 3
using the package 2to3, and then debugged as necessary.

The originally required Networkx == 1.11 is not supported in
Python 3, and had to be updated to the most recent version
(currently Networkx 3.2.1). The structure of the packaged
changed somewhat going from v1 -> 2 -> 3, and so in the
pyDAEDALUS scripts that use Networkx modules, modifications
were made as needed to ensure that the function with the
v3 Networkx is equivalent to the original code with the v1
package.

One important note: When calculating a minimum spanning tree
using Prim's algorithm, if there are multiple equivalently
minimal trees for the input geometry, Networkx v3 may return
a different tree than Networkx v1, presumably because of
a difference in the list/iterator output order. A different
but equivalently minimal spanning tree from Prim's algorithm
will still work with the rest of the pyDAEDALUS code to output
a foldable origami design, but the exact scaffold route, and
consequently the staple sequences, will be different. Thus,
this updated Python 3 and Networkx==3.2.1 version of pyDAEDALUS
may output different sequence designs for some geometries
than the original Python 2 and Networkx==1.11 version that was
originally released in 2022 and published in Parsons MF, Allan
MF, Li S et al (2023). The alternative sequence designs should
however not affect the ability of the object to fold, as they
are still derived from the same principle of a maximally
branched minimum spanning tree.